+++ resources/views/Admin/pembuatan-surat/surat-mati.blade.php
@@ -1,0 +1,20 @@
+<div class="mb-3">
+    <label for="nama_kematian" class="form-label"><b>Nama</b></label>
+    <input type="text" class="form-control" id="namaKematian" required>
+</div>
+<div class="mb-3">
+    <label for="jenisKelamin" class="form-label"><b>Jenis Kelamin</b></label>
+    <select class="form-select" id="jenisKelamin" aria-label="Jenis Kelamin">
+        <option selected>Pilih Jenis Kelamin</option>
+        <option value="laki-laki">Laki-laki</option>
+        <option value="perempuan">Perempuan</option>
+    </select>
+</div>
+<div class="mb-3">
+    <label for="alamatBapak" class="form-label"><b>Alamat</b></label>
+    <input type="text" class="form-control" id="alamatBapak" required>
+</div>
+<div class="mb-3">
+    <label for="tanggalLahirBapak" class="form-label"><b>Tanggal Kematian</b></label>
+    <input type="date" class="form-control" id="tanggalLahirBapak" required>
+</div>
